backend/pyproject.toml
[tool.poetry]
name = "cubiet"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
eth-abi = "^3.0.1"
eth-account = "^0.8.0"
rubiktwophase = "^1.1.1"
numpy = "^1.23.5"
opencv-python-headless = "^4.6.0.66"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
run = 'src.cubiet.main:main'
